requirements.txt
python-telegram-bot>=20.0
deep-translator>=1.10.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.25.1
beautifulsoup4>=4.9.3
httpx>=0.23.0
